runtime: ImageConfig, inclusind original image reference and changed storage schema
