--- utils/unpack_test.go
@@ -1,6 +1,0 @@
-package utils
-
-import "testing"
-
-func TestUnpack(t *testing.T) {
-}
